- Created a new Database and merged the cabins. - Modified the logic of onTap function. - Renamed the right cabin to cabin and right cabin repository to cabin repository - Deleted the unwanted files
