+++ .gitignore
@@ -42,3 +42,6 @@ app.*.map.json
 /android/app/debug
 /android/app/profile
 /android/app/release
+/pubspec.lock
+/android/package.json
+/android/package-lock.json
